+++ src/main/java/youth/week6/member/controller/MemberController.java
@@ -15,21 +15,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import youth.week6.argumentResolver.MemberId;
-import youth.week6.member.dto.request.LoginRequestDto;
-import youth.week6.member.dto.request.OrganizerJoinRequestDto;
-import youth.week6.member.dto.request.OrganizerMemberJoinRequestDto;
-import youth.week6.member.dto.request.ParticipantJoinRequestDto;
-import youth.week6.member.dto.request.ParticipantMemberJoinRequestDto;
-import youth.week6.member.dto.response.MemberDetailResponseDto;
-import youth.week6.member.dto.response.MemberJoinResponseDto;
+import youth.week6.member.controller.dto.request.LoginRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantMemberJoinRequestDto;
+import youth.week6.member.controller.dto.response.MemberDetailResponseDto;
+import youth.week6.member.controller.dto.response.MemberJoinResponseDto;
 import youth.week6.member.jwt.dto.TokenInfo;
 import youth.week6.member.controller.mapper.DtosToMemberDetailResponseDtoMapper;
 import youth.week6.member.controller.mapper.MemberJoinRequestDtoToJoinDtoMapper;
 import youth.week6.member.controller.mapper.OrganizerJoinRequestDtoToJoinDtoMapper;
 import youth.week6.member.controller.mapper.ParticipantJoinRequestDtoToJoinDtoMapper;
-import youth.week6.member.member.dto.MemberJoinDto;
-import youth.week6.member.organizer.dto.OrganizerJoinDto;
-import youth.week6.member.participant.dto.ParticipantJoinDto;
+import youth.week6.member.dto.MemberJoinDto;
+import youth.week6.member.dto.OrganizerJoinDto;
+import youth.week6.member.dto.ParticipantJoinDto;
 import youth.week6.member.service.LoginFacadeService;
 import youth.week6.member.service.MemberFacadeService;
 import youth.week6.member.service.dto.MemberDetailDto;
