+++ src/main/java/youth/week6/member/controller/mapper/MemberJoinRequestDtoToJoinDtoMapper.java
@@ -2,8 +2,8 @@ package youth.week6.member.controller.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import youth.week6.member.dto.request.MemberJoinRequestDto;
-import youth.week6.member.member.dto.MemberJoinDto;
+import youth.week6.member.controller.dto.request.MemberJoinRequestDto;
+import youth.week6.member.dto.MemberJoinDto;
 
 @Mapper(componentModel = "spring")
 public interface MemberJoinRequestDtoToJoinDtoMapper {
