+++ src/main/java/youth/week6/member/controller/mapper/OrganizerJoinRequestDtoToJoinDtoMapper.java
@@ -1,8 +1,8 @@
 package youth.week6.member.controller.mapper;
 
 import org.mapstruct.Mapper;
-import youth.week6.member.dto.request.OrganizerJoinRequestDto;
-import youth.week6.member.organizer.dto.OrganizerJoinDto;
+import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
+import youth.week6.member.dto.OrganizerJoinDto;
 
 @Mapper(componentModel = "spring")
 public interface OrganizerJoinRequestDtoToJoinDtoMapper {
