+++ src/main/java/youth/week6/member/controller/mapper/ParticipantJoinRequestDtoToJoinDtoMapper.java
@@ -1,8 +1,8 @@
 package youth.week6.member.controller.mapper;
 
 import org.mapstruct.Mapper;
-import youth.week6.member.dto.request.ParticipantJoinRequestDto;
-import youth.week6.member.participant.dto.ParticipantJoinDto;
+import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
+import youth.week6.member.dto.ParticipantJoinDto;
 
 @Mapper(componentModel = "spring")
 public interface ParticipantJoinRequestDtoToJoinDtoMapper {
