+++ src/test/java/youth/week6/member/member/entity/MembersTest.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import youth.week6.member.member.dto.MemberJoinDto;
+import youth.week6.member.dto.MemberJoinDto;
 
 class MembersTest {
